fix: 🧩 fixed the error caused by attempting to read a non-existent file.
